requirements.txt
# LogisticsAI Python Dependencies
# Core optimization algorithms and backend pipeline requirements

# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# HTTP Requests and API Integration
requests>=2.25.0
urllib3>=1.26.0

# Data Visualization and Analysis
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter Notebook for Analysis
jupyter>=1.0.0
ipykernel>=6.0.0
nbformat>=5.1.0

# Testing Framework
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0

# Code Quality and Formatting
black>=21.0.0
flake8>=3.9.0
isort>=5.9.0
mypy>=0.910

# Development and Debugging
ipdb>=0.13.0
python-dotenv>=0.19.0

# Performance Monitoring
memory-profiler>=0.60.0
line-profiler>=3.3.0

# Optional: Web Framework for API Deployment
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.1.0

# Optional: Advanced Optimization Libraries
# scikit-learn>=0.24.0  # For ML-based enhancements
# networkx>=2.5.0       # For graph-based algorithms
# numba>=0.53.0         # For JIT compilation optimization

# Development Environment
# Recommended for development but not strictly required
# jupyter-lab>=3.0.0
# jupyterlab-git>=0.30.0
